+++ index.html
@@ -7,7 +7,6 @@
   <link rel="stylesheet" href="css/normalize.css">
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
-  <script src="/js/main.js"></script>
   <script src="/functions.js"></script>
   <title>Кекстаграм</title>
 </head>
@@ -237,6 +236,7 @@
     </section>
   </template>
 
+  <script src="/js/main.js" type="module"></script>
 </body>
 
 </html>
