+++ js/data.js
@@ -1,0 +1,83 @@
+import { getRandomInt } from './util';
+
+const PHOTO_DESCRIPTIONS = [
+  'Солнечный закат над горами',
+  'Пустынный пляж с пальмами',
+  'Кошка на старом кресле',
+  'Дождливый город ночью',
+  'Лесная тропинка весной',
+  'Собака играет в парке',
+  'Зимний пейзаж с ёлками',
+  'Мост через тихую реку',
+  'Городской рынок с фруктами',
+  'Птица на ветке дерева',
+  'Океанские волны на закате',
+  'Старинный замок на холме',
+  'Дети катаются на велосипедах',
+  'Река среди осенних деревьев',
+  'Лаванда в солнечном поле',
+  'Кофе и книга на столе',
+  'Горный пейзаж с туманом',
+  'Ночной небоскрёб с огнями',
+  'Пикник на зелёной лужайке',
+  'Мотоцикл на пустой дороге',
+  'Рыба в прозрачной воде',
+  'Портрет улыбающейся девушки',
+  'Снежный город с огнями',
+  'Звёздное небо над пустыней',
+  'Яркие осенние листья на земле'
+];
+
+const DESCRIPTION_MESSAGE = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const NAMES = [
+  'Алексей',
+  'Иван',
+  'Дмитрий',
+  'Сергей',
+  'Михаил',
+  'Анна',
+  'Елена',
+  'Ольга',
+  'Мария',
+  'Татьяна'
+];
+
+//Функция генерации массива объектов описания фото
+const photos = function () {
+  const photosArr = [];
+  for (let i = 0; i < 25; i++) {
+    const commentsArr = [];
+    for (let j = 0; j < getRandomInt(0, 30); j++) {
+      commentsArr.push({
+        id: getRandomInt(0, 1000),
+        avatar: `img/avatar-${ getRandomInt(1, 6) }.svg`,
+        message: DESCRIPTION_MESSAGE[getRandomInt(0, DESCRIPTION_MESSAGE.length - 1)],
+        name: NAMES[getRandomInt(0, NAMES.length - 1)]
+      });
+    }
+    photosArr.push({
+      id: i,
+      url: `photos/${ i }.jpg`,
+      description: PHOTO_DESCRIPTIONS[i - 1],
+      likes: getRandomInt(15, 200),
+      comments: commentsArr
+    });
+  }
+  for (let i = photosArr.length - 1; i > 0; i--) {
+    const j = Math.floor(Math.random() * (i + 1));
+    [photosArr[i], photosArr[j]] = [photosArr[j], photosArr[i]];
+  }
+  return photosArr;
+};
+
+photos();
+
+console.log(photos());
